requirements.txt
diskcache~=5.6.3
httpx~=0.28.1
huggingface_hub~=0.35.0
ImageHash~=4.3.1
loguru~=0.7.3
numpy
pandas~=3.0.0
prometheus-client~=0.23.1
pyarrow~=19.0.1
PyYAML~=6.0.2
python-dotenv~=1.2.1
scikit-learn~=1.8.0
selectolax~=0.3.26
tqdm~=4.67.1
typer~=0.21.1
ultralytics~=8.3.218
transformers~=4.57.1
scipy~=1.17.0
torch~=2.10.0
torchvision~=0.25.0
pillow~=12.1.0
-e .
